Fix nonzero count on partitioned matrix

+++ internal/ceres/partitioned_matrix_view.h
@@ -98,15 +98,15 @@ class PartitionedMatrixViewBase {
   virtual void UpdateBlockDiagonalFtF(
       BlockSparseMatrix* block_diagonal) const = 0;
 
-  virtual int num_col_blocks_e() const = 0;
-  virtual int num_col_blocks_f() const = 0;
-  virtual int num_cols_e()       const = 0;
-  virtual int num_cols_f()       const = 0;
-  virtual int num_nonzeros_e()   const = 0;
-  virtual int num_nonzeros_f()   const = 0;
-  virtual int num_rows()         const = 0;
-  virtual int num_cols()         const = 0;
-  virtual int num_nonzeros()     const = 0;
+  virtual int num_col_blocks_e()   const = 0;
+  virtual int num_col_blocks_f()   const = 0;
+  virtual int num_cols_e()         const = 0;
+  virtual int num_cols_f()         const = 0;
+  virtual int64_t num_nonzeros_e() const = 0;
+  virtual int64_t num_nonzeros_f() const = 0;
+  virtual int num_rows()           const = 0;
+  virtual int num_cols()           const = 0;
+  virtual int num_nonzeros()       const = 0;
 
   static PartitionedMatrixViewBase* Create(const LinearSolver::Options& options,
                                            const BlockSparseMatrix& matrix);
@@ -132,8 +132,8 @@ class PartitionedMatrixView : public PartitionedMatrixViewBase {
   void UpdateBlockDiagonalFtF(BlockSparseMatrix* block_diagonal) const final;
   int num_col_blocks_e() const final { return num_col_blocks_e_;      }
   int num_col_blocks_f() const final { return num_col_blocks_f_;      }
-  int num_nonzeros_e()   const final { return num_nonzeros_e_;        }
-  int num_nonzeros_f()   const final { return num_nonzeros_f_;        }
+  int64_t num_nonzeros_e()   const final { return num_nonzeros_e_;        }
+  int64_t num_nonzeros_f()   const final { return num_nonzeros_f_;        }
   int num_cols_e()       const final { return num_cols_e_;            }
   int num_cols_f()       const final { return num_cols_f_;            }
   int num_rows()         const final { return matrix_.num_rows();     }
@@ -150,8 +150,8 @@ class PartitionedMatrixView : public PartitionedMatrixViewBase {
   int num_col_blocks_f_;
   int num_cols_e_;
   int num_cols_f_;
-  int num_nonzeros_e_;
-  int num_nonzeros_f_;
+  int64_t num_nonzeros_e_;
+  int64_t num_nonzeros_f_;
 };
 
 }  // namespace internal
